requirements.txt
ascii==3.6
asgiref==3.7.2
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0
django-admin-rangefilter==0.13.0
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-filter==23.5
django-jalali==6.0.1
django-jalali-date==1.1.0
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-location-field==2.7.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
idna==3.6
import-export==0.3.1
inflection==0.5.1
jdatetime==4.1.1
MarkupPy==1.14
mysqlclient==2.2.1
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
persian==0.5.0
persiantools==3.0.1
pillow==10.2.0
PyJWT==2.8.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.2.0
requests==2.31.0
setuptools==75.0.0
six==1.16.0
smsir-python==1.0.7
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
xlrd==2.0.1
xlwt==1.3.0
boto3
django-storages
django-import-export
